SDP updates after merging with 1.0.0 mainline
example_c6_env.sh: update license
examples/xml/to: remove junk file
hdl/devices/sdp_receive.hdl/sdp_receive_dma.vhd: fix active-flow-control role bug
hdl/primitives/Makefile: update to use new hdl-primitives.mk properly
hdl/primitives/ocpi/proptypes.cpp_gen: fix uchar_array property readback issue in sim
hdl/primitives/sdp/sdp_sim.vhd: update to use new verilog file output module
-- this avoids the need to patch isim
-- this worksaround modelsim "endfile" issue
hdl/primitives/util/util_pkg.vhd: add verilog file writer-with-flush
runtime/application/include/OcpiApplication.h: valgrind bug w/ errors thrown, port param support
runtime/application/src/OcpiApplication.cxx: ditto
runtime/application/src/ocpirun_main.cxx: remove duplicate option flag, try/catch cleanup
-- deployment support
runtime/dataplane/rdma_drivers/dma/src/DtDmaXfer.cxx: warning
runtime/hdl-support/src/HdlSimServer.cxx: share more code with LibraryManager etc.
runtime/hdl/src/HdlLSimDevice.cxx: ditto
runtime/library/include/OcpiLibraryManager.h: ditto
runtime/library/src/LibrarySimple.cxx: ditto
runtime/library/src/OcpiComponentLibrary.cxx: ditto
runtime/library/src/OcpiLibraryManager.cxx: ditto, and valgrind leak
runtime/library/src/OcpiLibraryAssembly.cxx: comment
runtime/util/assembly/include/OcpiUtilAssembly.h: another entry point for port params
runtime/util/assembly/src/OcpiUtilAssembly.cxx: ditto
tools/cdk/include/rcc/rcc-worker.mk: more rationalization of compiler flags, more needed.
hdl/platforms/isim/isim.mk: add non-pf
hdl/platforms/modelsim/modelsim.mk: ditto
hdl/primitives/util/flush_writer.vhd: add file-write-with-flush from verilog
hdl/primitives/util/flush_writer_internal.v: ditto

+++ runtime/dataplane/rdma_drivers/dma/src/DtDmaXfer.cxx
@@ -234,7 +234,7 @@ namespace OCPI {
 	  // A request to enable mapping to this bus address/size and return the physaddr
 	  if (ioctl(m_dmaFd, OCPI_CMD_REQUEST, &request))
 	    throw OU::Error("Can't establish remote DMA memory size %zu at 0x%" PRIx64
-			    "for DMA memory", (uint64_t)request.actual, ep.m_busAddr);
+			    "for DMA memory", (size_t)request.actual, ep.m_busAddr);
 	  ep.address = request.address;
 	  if (ep.m_holeOffset) {
 	    memset(&request, 0, sizeof(request));
